move monitors to graph_builder.go
